+++ CHANGES.rst
@@ -5,6 +5,8 @@ Release Notes
 Version 2.0.1.dev0 (in development)
 ====================================
 * (PENDING) Major changes to ambient condition recording and integration
+* Reverted AT106 get_mass_stable to 3 readings without delay (MettlerToledo base class implementation)
+* Added print of mass value to log!
 
 Version 2.0.0 (10/12/2024)
 ==========================